Scroll to element works with offset
Currently, the offset is determined with a hard pixel value
The navbar height is returning null in my javascript file. I believe
this is because I didn't actually set a 'height' attribute in my
navbar's CSS, the height is only determined by the content + padding.

The goal is to set the offset of the scroll to be equal to the height
of the navbar, whether in mobile, tablet or desktop configuration.